+++ components/script/layout_interface.rs
@@ -12,7 +12,6 @@ use euclid::point::Point2D;
 use euclid::rect::Rect;
 use gfx_traits::LayerId;
 use ipc_channel::ipc::{IpcReceiver, IpcSender};
-use libc::uintptr_t;
 use msg::compositor_msg::Epoch;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineId};
 use msg::constellation_msg::{WindowSizeData};
@@ -25,11 +24,12 @@ use std::any::Any;
 use std::sync::Arc;
 use std::sync::mpsc::{Receiver, Sender, channel};
 use string_cache::Atom;
-use style::animation::PropertyAnimation;
 use style::stylesheets::Stylesheet;
 use url::Url;
 use util::ipc::OptionalOpaqueIpcSender;
 
+pub use style::animation::Animation;
+pub use style::context::ReflowGoal;
 pub use dom::node::TrustedNodeAddress;
 
 /// Asynchronous messages that script can send to layout.
@@ -138,15 +138,6 @@ impl OffsetParentResponse {
     }
 }
 
-/// Why we're doing reflow.
-#[derive(PartialEq, Copy, Clone, Debug)]
-pub enum ReflowGoal {
-    /// We're reflowing in order to send a display list to the screen.
-    ForDisplay,
-    /// We're reflowing in order to satisfy a script query. No display list will be created.
-    ForScriptQuery,
-}
-
 /// Any query to perform with this reflow.
 #[derive(PartialEq)]
 pub enum ReflowQueryType {
@@ -223,30 +214,6 @@ impl ScriptLayoutChan for OpaqueScriptLayoutChannel {
     fn receiver(self) -> Receiver<Msg> {
         let OpaqueScriptLayoutChannel((_, receiver)) = self;
         *receiver.downcast::<Receiver<Msg>>().unwrap()
-    }
-}
-
-/// Type of an opaque node.
-pub type OpaqueNode = uintptr_t;
-
-/// State relating to an animation.
-#[derive(Clone)]
-pub struct Animation {
-    /// An opaque reference to the DOM node participating in the animation.
-    pub node: OpaqueNode,
-    /// A description of the property animation that is occurring.
-    pub property_animation: PropertyAnimation,
-    /// The start time of the animation, as returned by `time::precise_time_s()`.
-    pub start_time: f64,
-    /// The end time of the animation, as returned by `time::precise_time_s()`.
-    pub end_time: f64,
-}
-
-impl Animation {
-    /// Returns the duration of this animation in seconds.
-    #[inline]
-    pub fn duration(&self) -> f64 {
-        self.end_time - self.start_time
     }
 }
